edits

+++ src/strategies/pct_strategy.cc
@@ -16,7 +16,7 @@ namespace coyote
 	{
 	}
 
-	int PCTStrategy::next_operation(Operations& operations, size_t current)
+	size_t PCTStrategy::next_operation(Operations& operations, size_t current)
 	{
 		set_new_operation_priorities(operations, current);
 		try_deprioritize_operation_with_highest_priority(operations);
